requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
Django==2.2.16
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pyparsing==3.0.9
pytz==2022.2.1
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.11